+++ Python Library/max_pool_layer.py
@@ -19,7 +19,7 @@ class max_pool_layer(theano_layer):
     if dropout_prob > 0.0:
       if batch_size == None:
         batch_size = self.num_training_examples
-      self.dropout_mask  = trng.binomial(n = 1, p = 1 - dropout_prob, size = np.insert(input_shape, 0, batch_size)) / dropout_prob
+      self.dropout_mask  = trng.binomial(n = 1, p = 1 - dropout_prob, size = np.insert(input_shape, 0, batch_size), dtype = 'float32') / dropout_prob
       self.masked_output = T.signal.downsample.max_pool_2d(input = X_var * self.dropout_mask , ds = (stride, stride), ignore_border = False)
     else:
       self.masked_output = downsample.max_pool_2d(input = X_var, ds = (stride, stride), ignore_border = False)
